pytorch/requirements.txt
fastapi
uvicorn
numpy
pillow
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchvision==0.18.1+cpu
python-multipart